+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - feature/1200942872266138 - Add HasValueResolverInterface
 - feature/1201104104768447 - Add filterByValueType interface and trait (filters PHP types inside mixed collection)
 - feature/1201127836211937 - Add unique and cast methods on IterableHelper. Add ComparisonInterface
+- feature/1201194348713239 - Add LDL_TYPE_SCALAR and filterByValueType support in IterableHelper
 
 ### Changed